+++ templates/archive.php
@@ -60,4 +60,4 @@ pp_base_template(function () {
         ]);
     }
 
-});
+}, ['hero', 'grid', 'section']);