+++ templates/search.php
@@ -53,4 +53,4 @@ pp_base_template(function () {
         ]);
     }
 
-});
+}, ['hero', 'grid', 'section']);